Big commit--use personal content provider, correctly use cursor adapter, correctly display scrollable fragment with listview and scrollable other views Load REST data to sql db, join tables

+++ app/src/main/java/com/example/ahmed/popularmovies/rest/MovieReviewsFetchService.java
@@ -10,6 +10,6 @@ import retrofit.http.Path;
 public interface MovieReviewsFetchService {
 
     @GET("/3/movie/{id}/reviews?api_key=" + API_KEY.API_KEY)
-    Call<MovieReviews> reviewsList(@Path("id") Integer id);
+    Call<MovieReviews> reviewsList(@Path("id") long id);
 
 }
